Remove "stderr"
I find it annoying because it adds a lot of noise to the godoc and
autocomplete, with no real value.

We use it in only four places in Desk, and when you use `return
httperr.ErrFoo` the stack traces are wrong in Sentry, too :-/ Besides,
usually you want a more descriptive message anyway like `not allowed to
access inbox %v` or some such.