Rename to guru
I think that "httperr" is a misnomer, and a confusing one. All this
package does is add an error code to an error message, which is similar
to for example MySQL:

    Error 1317: Query execution was interrupted
    Error 1451: Cannot delete or update a parent row: a foreign key constraint fails

And so forth.

For most of our applications, it makes sense to use error codes
identical to HTTP error codes, because *most* of the time we're using
code in a HTTP context, but this is not in any way mandatory. We can
define our own.

Renaming it should solve the confusion that this package is somehow tied
to the HTTP protocol, which it isn't. There are no real code changes
required for this. Using an error code of "4" works already. It's jut a
matter of naming and phrasing the documentation. It should also make it
more natural to use this package in non-HTTP contexts.

I chose `guru` because it is short, sounds nice, is slightly humorous,
and because I was reading about the Amiga the other day.

Migration:

    find . -name '*.go' -exec sed -i \
        -e 's!github.com/teamwork/httperr!github.com/teamwork/guru!g;' \
        -e 's/httperr\.WithStatus/guru.WithCode/g;' \
        -e 's/httperr\.StatusCode/guru.Code/g;' \
        -e 's/httperr\.UserError/guru.HTTPUserError/g;' \
        -e 's/httperr\./guru./g;' \
        {} \;

    find . -name '*.go' -exec goimports -w {} \;